notification paging support

+++ Posts/API/dao/src/main/java/dido/auntaccount/dto/NotificationListDTO.java
@@ -4,23 +4,23 @@ import java.util.List;
 
 public class NotificationListDTO {
 
-    private Integer total;
+    private long total;
     private List<NotificationDTO> rows;
 
     public NotificationListDTO() {
 
     }
 
-    public NotificationListDTO(Integer total, List<NotificationDTO> rows) {
+    public NotificationListDTO(long total, List<NotificationDTO> rows) {
         this.total = total;
         this.rows = rows;
     }
 
-    public Integer getTotal() {
+    public long getTotal() {
         return total;
     }
 
-    public void setTotal(Integer total) {
+    public void setTotal(long total) {
         this.total = total;
     }
